Add `vm` mid velocity to Moves.
This means the nominal target velocity `v` is not modified, with `vm`
containing the actual achieved middle velocity. This helps remove spurious
extra `Fxxx` arguments in output setting the velocity to the extimated
achieved velocity, allowing the printer to go up to the target velocity if it
can do faster than extimated.