Tidy up gcodgen.py Move stuff.
Change the Move.vm attribute to a cached_property instead of calculating it in
Move.fixv().

Add Move.setv() for setting v0 and v1 that clears all cached_propertys that
depend on them, and use it in Move.__new__(), Move._replace(), and
Move.fixv().

Rename Move.set() to Move.change(), and use it everywhere instead of
Move._replace().

Rename functions isneq() to isneareq() and islne() to isnearle(). Add
getnear() function to round values to 6 decimals and turn -0.0 into 0.0, which
is "near" enough to pass the isneareq() and isnearle() tests. Use getnear()
instead of round() everywhere.

Tidy Move.join() and Move.split() a bit.