Another significant refactor of gcodegen.py.
This makes `-O` capable of merging hops and retracts/restores, changes `-V` to
`-V=[0-3]` with three levels of velocity fixing, adds `-S` for partitioning
moves into accel/cruise/decel phases independently of `-P`, and adds a
";command_line: <cmdline>" comment to startcode so we can see what the cmdline
used was.

It tidies up the architecture for how adding/finalizing/post-processing is done.