Tidy misc things in gcodegen.py.
Change isneareq() and isnearle() functions to take "n" number of decimals
instead of "e" error to make it consistent with getnear().

Add islayer(), ismove(), etc helper methods for checking the type of cmds and
use them eveywhere.

Make inccmd() always resetstate() before performing actions so that state
variables always reflects the results of the last command whether it was a
move or any kind of gcode.

Fix inccmd() for G4 cmds to correctly call self.inct().