Significant refactor of conv-gcode.py for new gcodegen.py architecture.
This leverages of iterstate() to calculate the filters, inc() methods to
update state from commands, and fadd() methods to finalize the output fan
and pause commands depending on the cmdline args.

The runCode() method was renamed addCode() to compliment getCode(), and the
line adding mechanism was renamed to padd() (parse-add).