Tidy up layer and retraction code for cleaner conv-gcode.py output.
In gcodegen.py make resetlayer() always set a default base layer, and remove a
bunch of special handling for the case when self.layer was None. Make the
`self.w` and `self.r` properties return the previous draw's attributes before
falling back to the layer defaults. Make `flayer()` format the zero layer as
a ";preExtrude;<h>` line. Stop fmove() adding 'extrude_ratio` comments. Make
`endlayer()` not do a `hopup()`. Swaped `move()` and `draw()` again so move is
the implementation and `draw()` is calls it with `r` defaulting to the value
used for the previous draw. Added a `pe` argument to `retract()` and
`restore()` which is the target pressue to retract/restore to.

In conv-gcode.py remove code seting an initial base layer and emiting
`preExtrude` comments.

In gxgen tidy it up and make it use `git sdiff` to show the changes.