Fix restore() so that it cannot be mistaken for a retract().
It's possible that the model sees vestigial nozzle pressure before doing a
restore() which results in restore() generating a retraction to zero the
pressure. With `dynret` turned on this gets recognized as and turned into a
fully blown retraction instead of a restore. We need to make sure restore()
always generates a small restore to ensure this doesn't happen.