Make bead pressure scale with w only up to 2x the nozzle diameter.
This means the bead backpressure increases with line width upto 2x the nozzle
width and then "flatlines". This generally reflects observations from the
SpStartStopTest testing. The theory is that once the bead gets past the edge
of the nozzle it "bulges up" and doesn't have to press out which reduces the
pressure.

This was implemented with a few refactors to centralize where the pressure
calculations are done.

Add Pb(), Pn(), and Pe() methods for calculating bead, nozzle, and extruder
pressures from bead diameter and nozzle flow velocity. The Pb() method limits
the bead width for calculating backpressure with Kb to 2x the nozzle diameter.

Add pn and pb properties for the current nozzle and bead pressure.

Use these methods and properties wherever pressure is needed, replacing any
duplicated calculations. Remove the now redundant `_calc_pe()` method which is
replaced by the new Pe() method.

Also tidy up the dynret code P=0.8 scaling when dynext is on, making it
simpler.