Adjust gcodetest.py spiral tests for better warmup/cooldown.
This changes the warmup and cooldown draws from vl=1mm/s to vl=5mm/s and
adjusts the retract/restore distances to try and endure they always start with
the right pressure.

It also tweaks the dial start line drawing so we only draw it at the start of
the outer most tests.