In gcodegen.py improve `-P` dynext functionality.
Make dynext use `_getNextPeEb()` to get the target pressure the same as used
for dynret retract/restore adjustments.

Make `_getNextPeEb()` always look ahead up to the next restore/retract instead of
just upto the next non-draw or only one draw if dynext is on. Make it
calculate eb using a line-average the same as for ve and db instead of just
the next draw.