Make Move's `r` and `de` independent and improve logging.
This makes a Move's `r` extrusion ratio reflect the original intent which can
be independent of the extrusion `de` which can include pressure advance
adjustments. By default `r` is calculated based on `de` for moves and draws,
or set to the retract/restore distance for other types of moves. When
dynamically adjusting extrusions with `_replace()` the old `r` intent is
preserved. A new `er` property gives the adjusted extrusion ratio for the
actual `de`, making it possible to compare the intended ratio with the
adjusted extrusion ratio.

The verbose logging output has been changed to be more char-efficient
formatting numbers and more clearly show the pressure advance extrusion ratio
changes from the original intent (as indicated by the `r` attribute) to the
extruded ratio (as indicated by the `de` attribute), to the nozzle output (as
calculated by the simulation).

Also fix `canjoin()` to correctly check that the join candidates have the same
line attributes.