Always use latest move's h,w,r for extrusion volumes and tidy names.
In conv-gcode.py rename target width attribute `_w` to `wt` and always use hl
for parsed non-hop cmds at h=0.

in gcodegen.py rename properties w and r to wl and rl to match hl and make
them all default to the latest move, falling back to the layer defaults. Make
incmove() update self.m at the start not end so that the processing of the
move can use it, in particular make sure hl,wl,rl are for the current move.