Improve classification of move types and fix gcodetest.py.
This changes the classification of moves into non-overlapping isstop, isshift,
ishop, isadjust categories based on the horizontal, vertical, and extrude
distances relative to their max-speeds. This means they can all include dl,
dz, de movement and are classified by which is the dominant one.

Also change the r values to allow negative values, mainly for retractions. The
initial defaults are -1 or 1 for retract/restore adjustments, otherwise use
r=et/el. The el and et calculations have been cleaned up a bit to reflect the
classification and r chages.

Also initialize the stats with arrays and a zero count. This ensures they are
iitialized to the right size and report correctly even when no values have
been added.

Add some more assertions and fix a few for the r changes.

Fix gcodetest.py for the self.w -> self.wl change in gcodegen.py.